Made Wrapper more robust by requiring the arguments it must have to execute. Fixed a bug where the order was always set to be original.

+++ dt-impact-tracer/src/edu/washington/cs/dt/impact/tools/FileTools.java
@@ -98,8 +98,10 @@ public class FileTools {
     // deletes the files in filesToDelete.
     // used to clear the environment of any temp files created by tests
     public static void clearEnv(List<String> filesToDelete) {
-        for (String s : filesToDelete) {
-            new File(s).delete();
+        if (filesToDelete != null) {
+            for (String s : filesToDelete) {
+                new File(s).delete();
+            }
         }
     }
 }
